# 07_api_design/openapi.yaml
openapi: 3.0.3
info:
  title: Atelier API
  version: 1.0.0
  description: |
    Atelier - AI 社員常駐型開発プロジェクト統合管理 SaaS の API。

    ★ API 契約凍結: 2026-05-20T13:30:00+09:00 ★
    以後の変更は ADR + 影響解析 + 承認フロー必須。

    認証 4 系統:
    - 一般 JWT (Supabase Auth)
    - クライアント別 JWT (role=client_portal, project_id クレーム限定・R-T08 対応)
    - MCP Token
    - Service Role (バックエンド内部 + Bridge worker)
  contact:
    name: ENGINE BASE
    email: info@engine-base.com
  license:
    name: Proprietary
servers:
  - url: https://atelier.app/api/v1
    description: 本番
  - url: https://*.preview.atelier.app/api/v1
    description: Preview Deploy
  - url: http://localhost:8000/api/v1
    description: ローカル開発
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bridgeToken:
      type: apiKey
      in: header
      name: X-Bridge-Token
    clientPortalAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHENTICATED
                - INVALID_CREDENTIALS
                - INVALID_BRIDGE_TOKEN
                - INVALID_INVITATION_TOKEN
                - FORBIDDEN
                - ACCOUNT_LOCKED
                - ACCOUNT_DELETED
                - NOT_FOUND
                - CONFLICT
                - INVALID_LIFECYCLE_STAGE
                - DEPENDENCIES_NOT_MET
                - ALREADY_DECIDED
                - EMAIL_ALREADY_REGISTERED
                - EMAIL_IN_DELETION_GRACE
                - OWNER_OF_WORKSPACE_WITH_OTHERS
                - INVITATION_EXPIRED
                - VALIDATION_ERROR
                - CONSENT_REQUIRED
                - RATE_LIMITED
                - BYOK_QUOTA_EXCEEDED
                - INTERNAL_ERROR
                - BRIDGE_OFFLINE
                - PARALLEL_LIMIT_REACHED
                - F_CTX01_TIMEOUT
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
            request_id: { type: string, format: uuid }
            trace_id: { type: string }
...
    ClientInvitation:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        email: { type: string, format: email }
        scopes: { type: array, items: { type: string } }
        expires_at: { type: string, format: date-time }
        used_at: { type: string, format: date-time, nullable: true }
        revoked_at: { type: string, format: date-time, nullable: true }
        client_display_name: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ClientInvitationWithToken:
      allOf:
        - { $ref: '#/components/schemas/ClientInvitation' }
        - type: object
          properties:
            token: { type: string, description: 生成時のみ返る raw token (招待 URL 用) }
...
    ChatThread:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        ai_employee_id: { type: string, format: uuid }
        title: { type: string, nullable: true }
        archived: { type: boolean }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        thread_id: { type: string, format: uuid }
        role: { type: string, enum: [user, assistant, system, tool] }
        content: { type: string }
        parent_message_id: { type: string, format: uuid, nullable: true }
        token_count: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        target_type: { type: string, enum: [workflow_output, mock, task, acceptance_criteria] }
        target_id: { type: string, format: uuid }
        target_element_id: { type: string, nullable: true }
        author_user_id: { type: string, format: uuid, nullable: true }
        author_invitation_id: { type: string, format: uuid, nullable: true }
        content: { type: string }
        status: { type: string, enum: [open, resolved, deleted] }
        parent_comment_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /client-invitations:
    get:
      tags: [client-invitations]
      summary: クライアント招待一覧
      x-implementation-path: apps/api/src/routes/client_invitations/__init__.py::list_invitations
      x-screen-ids: [S-L01]
      parameters:
        - { name: project_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ClientInvitation' } }
    post:
      tags: [client-invitations]
      summary: クライアント招待作成
      x-implementation-path: apps/api/src/routes/client_invitations/__init__.py::create_invitation
      x-screen-ids: [S-L01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, email]
              properties:
                project_id: { type: string, format: uuid }
                email: { type: string, format: email }
                scopes: { type: array, items: { type: string } }
                client_display_name: { type: string, maxLength: 100 }
                ttl_days: { type: integer, minimum: 1, maximum: 30, default: 7 }
      responses:
        '201':
          description: 作成成功（raw token を含む）
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ClientInvitationWithToken' }
  /client-invitations/{invitation_id}:
    get:
      tags: [client-invitations]
      summary: クライアント招待詳細
      x-implementation-path: apps/api/src/routes/client_invitations/__init__.py::get_invitation
      x-screen-ids: [S-L01]
      parameters:
        - { name: invitation_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ClientInvitation' }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /client-invitations/{invitation_id}/revoke:
    post:
      tags: [client-invitations]
      summary: クライアント招待失効
      x-implementation-path: apps/api/src/routes/client_invitations/__init__.py::revoke_invitation
      x-screen-ids: [S-L01]
      parameters:
        - { name: invitation_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 失効成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ClientInvitation' }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 既に失効済, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ==========================================================================
  # workflow (工程一覧・遷移) / outputs (成果物)
  # ==========================================================================
...
  /comments:
    get:
      tags: [comments]
      summary: コメント一覧（対象指定）
      x-implementation-path: apps/api/src/routes/comments/__init__.py::list_comments
      x-screen-ids: [S-G01, S-H01, S-L03]
      parameters:
        - { name: target_type, in: query, required: true, schema: { type: string, enum: [workflow_output, mock, task, acceptance_criteria] } }
        - { name: target_id, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Comment' } }
    post:
      tags: [comments]
      summary: コメント作成
      x-implementation-path: apps/api/src/routes/comments/__init__.py::create_comment
      x-screen-ids: [S-G01, S-H01, S-L03]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_type, target_id, content]
              properties:
                target_type: { type: string, enum: [workflow_output, mock, task, acceptance_criteria] }
                target_id: { type: string, format: uuid }
                content: { type: string, minLength: 1, maxLength: 10000 }
                target_element_id: { type: string, maxLength: 200 }
                parent_comment_id: { type: string, format: uuid }
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Comment' }
        '403': { description: 対象に対する権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /comments/{comment_id}:
    parameters:
      - { name: comment_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [comments]
      summary: コメント詳細
      x-implementation-path: apps/api/src/routes/comments/__init__.py::get_comment
      x-screen-ids: [S-G01, S-H01, S-L03]
      responses:
        '200':
          description: 詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Comment' }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [comments]
      summary: コメント編集・解決
      x-implementation-path: apps/api/src/routes/comments/__init__.py::update_comment
      x-screen-ids: [S-G01, S-H01, S-L03]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string, minLength: 1, maxLength: 10000 }
                status: { type: string, enum: [open, resolved] }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Comment' }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [comments]
      summary: コメント削除（論理）
      x-implementation-path: apps/api/src/routes/comments/__init__.py::delete_comment
      x-screen-ids: [S-G01, S-H01, S-L03]
      responses:
        '204': { description: 削除成功 }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ==========================================================================
  # chat (スレッド CRUD)
  # ==========================================================================
  /chat/threads:
    get:
      tags: [chat]
      summary: チャットスレッド一覧
      x-implementation-path: apps/api/src/routes/chat/__init__.py::list_threads
      x-screen-ids: [S-D01]
      parameters:
        - { name: project_id, in: query, schema: { type: string, format: uuid } }
        - { name: include_archived, in: query, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ChatThread' } }
    post:
      tags: [chat]
      summary: チャットスレッド作成
      x-implementation-path: apps/api/src/routes/chat/__init__.py::create_thread
      x-screen-ids: [S-D01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, ai_employee_id]
              properties:
                project_id: { type: string, format: uuid }
                ai_employee_id: { type: string, format: uuid }
                title: { type: string, maxLength: 200 }
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ChatThread' }
  /chat/threads/{thread_id}:
    parameters:
      - { name: thread_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [chat]
      summary: チャットスレッド詳細
      x-implementation-path: apps/api/src/routes/chat/__init__.py::get_thread
      x-screen-ids: [S-D01]
      responses:
        '200':
          description: 詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ChatThread' }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [chat]
      summary: チャットスレッド更新
      x-implementation-path: apps/api/src/routes/chat/__init__.py::update_thread
      x-screen-ids: [S-D01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200 }
                archived: { type: boolean }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ChatThread' }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [chat]
      summary: チャットスレッド削除（論理）
      x-implementation-path: apps/api/src/routes/chat/__init__.py::delete_thread
      x-screen-ids: [S-D01]
      responses:
        '204': { description: 削除成功 }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /chat/threads/{thread_id}/messages:
    parameters:
      - { name: thread_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [chat]
      summary: チャットメッセージ一覧
      x-implementation-path: apps/api/src/routes/chat/__init__.py::list_messages
      x-screen-ids: [S-D01]
      responses:
        '200':
          description: 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ChatMessage' } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [chat]
      summary: チャットメッセージ送信（即時 / ユーザー発話）
      x-implementation-path: apps/api/src/routes/chat/__init__.py::create_message
      x-screen-ids: [S-D01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 100000 }
      responses:
        '201':
          description: 送信成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ChatMessage' }
        '403': { description: 投稿権限なし (viewer 等), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ==========================================================================
  # admin (運営)
  # ==========================================================================
...